+++ src/data/elements.js
@@ -1,4 +1,4 @@
-import { $ } from "./utils.js";
+import { $ } from "../utils.js";
 
 export const $result = $("#result");
 export const $userInput = $("#user-input");
